rework pluralize stub question to not ask for 3 stubs
Asking for more than one stub runs completely counter to the teaching that stubs should not require thought, and that the body of any stub that produces a type should be the same.

It isn't until later than we teach using the parameter as the body of the stub. There's no need for that wrinkle in module 1b.